requirements.txt
streamlit>=1.30.0
openai>=1.12.0
linkup-sdk>=0.1.0
python-dotenv>=1.0.0
groq>=0.6.0
